mcp/demo/requirements.txt
fastmcp>=2.9.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
